Try casting to int

+++ cpp/dolfinx/common/MPI.cpp
@@ -97,7 +97,7 @@ dolfinx::MPI::compute_graph_edges_pcx(MPI_Comm comm, std::span<const int> edges)
   spdlog::info(
       "Computing communication graph edges (using PCX algorithm). Number "
       "of input edges: {}",
-      edges.size());
+      static_cast<int>(edges.size()));
 
   // Build array with '0' for no outedge and '1' for an outedge for each
   // rank
@@ -153,7 +153,7 @@ dolfinx::MPI::compute_graph_edges_pcx(MPI_Comm comm, std::span<const int> edges)
 
   spdlog::info("Finished graph edge discovery using PCX algorithm. Number "
                "of discovered edges {}",
-               other_ranks.size());
+               static_cast<int>(other_ranks.size()));
 
   return other_ranks;
 }
@@ -164,7 +164,7 @@ dolfinx::MPI::compute_graph_edges_nbx(MPI_Comm comm, std::span<const int> edges)
   spdlog::info(
       "Computing communication graph edges (using NBX algorithm). Number "
       "of input edges: {}",
-      edges.size());
+      static_cast<int>(edges.size()));
 
   // Start non-blocking synchronised send
   std::vector<MPI_Request> send_requests(edges.size());
@@ -234,7 +234,7 @@ dolfinx::MPI::compute_graph_edges_nbx(MPI_Comm comm, std::span<const int> edges)
 
   spdlog::info("Finished graph edge discovery using NBX algorithm. Number "
                "of discovered edges {}",
-               other_ranks.size());
+               static_cast<int>(other_ranks.size()));
 
   return other_ranks;
 }
